+++ src/records.ts
@@ -222,14 +222,14 @@ export interface SubdivisionsRecord {
 export interface TraitsRecord {
   /**
    * The autonomous system number associated with the IP address. This value is
-   * only set when using the City or Insights web service or the Enterprise
-   * database.
+   * only set when using the City Plus or Insights web service or the
+   * Enterprise database.
    */
   readonly autonomousSystemNumber?: number;
   /**
    * The organization associated with the registered autonomous system number
-   * for the IP address. This value is only set when using the City or Insights
-   * web service or the Enterprise database.
+   * for the IP address. This value is only set when using the City Plus or
+   * Insights web service or the Enterprise database.
    */
   readonly autonomousSystemOrganization?: string;
   /**
@@ -241,8 +241,8 @@ export interface TraitsRecord {
   /**
    * The second level domain associated with the IP address. This will be
    * something like "example.com" or "example.co.uk", not "foo.example.com".
-   * This value is only set when using the City or Insights web service or the
-   * Enterprise database.
+   * This value is only set when using the City Plus or Insights web service or
+   * the Enterprise database.
    */
   readonly domain?: string;
   /**
@@ -309,29 +309,30 @@ export interface TraitsRecord {
   network?: string;
   /**
    * The name of the ISP associated with the IP address. This value is only set
-   * when using the City or Insights web service or the Enterprise database.
+   * when using the City Plus or Insights web service or the Enterprise
+   * database.
    */
   readonly isp?: string;
   /**
    * The mobile country code (MCC) associated with the IP address and ISP.
    * See https://en.wikipedia.org/wiki/Mobile_country_code.
    *
-   * This value is only set when using the City or Insights web service or the
-   * Enterprise database.
+   * This value is only set when using the City Plus or Insights web service or
+   * the Enterprise database.
    */
   readonly mobileCountryCode?: string;
   /**
    * The mobile network code (MNC) associated with the IP address and ISP.
    * See https://en.wikipedia.org/wiki/Mobile_country_code.
    *
-   * This value is only set when using the City or Insights web service or the
-   * Enterprise database.
+   * This value is only set when using the City Plus or Insights web service or
+   * the Enterprise database.
    */
   readonly mobileNetworkCode?: string;
   /**
    * The name of the organization associated with the IP address. This value is
-   * only set when using the City or Insights web service or the Enterprise
-   * database.
+   * only set when using the City Plus or Insights web service or the
+   * Enterprise database.
    */
   readonly organization?: string;
   /**
@@ -349,8 +350,8 @@ export interface TraitsRecord {
    */
   readonly userCount?: number;
   /**
-   *  The user type associated with the IP address. This value is only set when
-   *  using the City or Insights web service or the Enterprise database.
+   * The user type associated with the IP address. This value is only set when
+   * using the City Plus or Insights web service or the Enterprise database.
    */
   readonly userType?:
     | 'business'